pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "protein-llm"
version = "0.1.0"
description = "Post-training LLMs for protein understanding"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.5.0",
    "transformers>=4.51.0",
    "peft>=0.10.0",
    "bitsandbytes>=0.43.0",
    "fair-esm>=2.0.0",
    "biopython>=1.86",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb",
    "tensorboard",
    "datasets>=2.18.0",
    "scipy",
    "pandas",
    "numpy<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "ruff",
    "mypy",
]
training = [
    "deepspeed>=0.14.0",
    "ray[default]>=2.10.0",
]
verl = [
    "verl>=0.2.0",
    "vllm>=0.6.0",
]

[project.scripts]
protein-train = "scripts.train:main"
protein-eval = "scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
